Localization: pt-BR + UI
Added the UI texts on localization file and changed the code to get the text values from there, making possible to translate to other languages.
Translated to brazilian portuguese all the configs, items, keybinds and popups, also translated the UI using the new way of getting the text from the localization file.